requirements.txt
pdfplumber>=0.10.0
pydantic>=2.0.0
pandas>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.27.0
python-dateutil>=2.8.0
pillow>=9.0.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2

